+++ NEWS.md
@@ -2,7 +2,7 @@
 
 # development version
 
-* Add Rower 2017 tacrolimus ([doi:10.1136/bmjpo-2017-000147](https://doi.org/10.1136/bmjpo-2017-000147)) -- paediatric heart transplant recipients within the first 6 weeks post-transplant.
+* Add Park 2014 triflusal ([doi:10.1186/2050-6511-15-75](https://doi.org/10.1186/2050-6511-15-75)) -- 34 healthy Korean adult male volunteers (one-compartment HTB metabolite popPK with binary probability PD for inhibition of platelet aggregation).
 * Add Padoin 1998 cephalexin ([doi:10.1128/aac.42.6.1463](https://doi.org/10.1128/aac.42.6.1463)) -- preclinical male Wistar rats with a cephalexin / quinapril oral coadministration DDI (ratifies new `CONMED_QPRL_ORAL` covariate canonical).
 * Add van Rongen 2018 metformin ([doi:10.1007/s40272-018-0293-1](https://doi.org/10.1007/s40272-018-0293-1)) -- overweight and obese Caucasian adolescents.
 * Add Tarning 2012 artemether and dihydroartemisinin ([doi:10.1186/1475-2875-11-293](https://doi.org/10.1186/1475-2875-11-293)) -- pregnant women with uncomplicated Plasmodium falciparum malaria in Uganda (joint parent + DHA popPK with zero-order dissolution and 6-compartment transit absorption).